+++ packages/core-js/features/weak-set/add-all.js
@@ -1,0 +1,4 @@
+require('../../modules/es.weak-set');
+require('../../modules/esnext.weak-set.add-all');
+
+module.exports = require('../../internals/entry-unbind')('WeakSet', 'addAll');