+++ packages/core-js/modules/esnext.weak-set.delete-all.js
@@ -1,9 +1,10 @@
 'use strict';
 var collectionDeleteAll = require('../internals/collection-delete-all');
+var IS_PURE = require('../internals/is-pure');
 
-// `Set.prototype.deleteAll` method
+// `WeakSet.prototype.deleteAll` method
 // https://github.com/tc39/collection-methods
-require('../internals/export')({ target: 'Set', proto: true, real: true, forced: require('../internals/is-pure') }, {
+require('../internals/export')({ target: 'WeakSet', proto: true, real: true, forced: IS_PURE }, {
   deleteAll: function deleteAll(/* ...elements */) {
     return collectionDeleteAll.apply(this, arguments);
   }